App/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cleo==0.8.1
click==8.1.3
clikit==0.6.2
colorama==0.4.6
contourpy==1.0.6
crashtest==0.3.1
cycler==0.11.0
dataclasses==0.6
dcmrtstruct2nii==2
debugpy==1.6.3
decorator==4.4.2
dicom2nifti==2.4.5
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.5.0
entrypoints==0.4
executing==1.2.0
fire==0.4.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
ignite==1.1.0
imageio==2.22.4
imageio-ffmpeg==0.4.7
ipykernel==6.17.0
ipython==8.6.0
ipywidgets==8.0.2
jedi==0.18.1
joblib==1.2.0
jsonargparse==4.18.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lightning-utilities==0.3.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
monai==0.7.0
moviepy==1.0.3
multidict==6.0.2
nest-asyncio==1.5.6
networkx==2.8.8
nibabel==4.0.2
numpy==1.23.4
opencv-python==4.6.0.66
openpyxl
packaging==21.3
pandas==1.5.1
parso==0.8.3
pastel==0.2.1
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.3.0
proglog==0.1.10
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.3
pure-eval==0.2.2
pyautogui
pydicom==2.3.0
Pygments==2.13.0
pylev==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-gdcm==3.0.19
pytorch-ignite==0.4.5
pytorch-lightning==1.8.3
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rt-utils==1.2.5
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
SimpleITK==2.2.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.0
tensorboardX==2.5
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.9.1+cu111
torchaudio==0.9.1
torchmetrics==0.10.3
torchvision==0.10.1+cu111
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.13.5
wcwidth==0.2.5
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.8.1
